refactor: capitlize component file names, update component index
